Implement constraints infrastructure without relying on `pre_randomize()`
The previous implementation blocked users from adding code to `pre_randomize()` in classes where `constraints_utils` was expanded. By shifting the code that handles constraints to a nested `rand` object, we allow its `pre_randomize()` to be called automatically. We also move non-trivial code outside of the macro, which is a win for readability.